# docs/swagger.yaml
basePath: /api
definitions:
...
  main.Hedgehog:
    description: Information about a hedgehog in the rescue center
    properties:
      area:
        $ref: '#/definitions/main.Area'
      area_id:
        example: 1
        type: integer
      arrival_date:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      description:
        example: Riccio trovato nel giardino
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Spillo
        type: string
      release_date:
        example: "2024-07-28T10:30:00Z"
        format: date-time
        type: string
      status:
        enum:
        - in_care
        - recovered
        - deceased
        - released
        example: in_care
        type: string
      therapies:
        items:
          $ref: '#/definitions/main.Therapy'
        type: array
      updated_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      weight_records:
        items:
          $ref: '#/definitions/main.WeightRecord'
        type: array
    type: object
...
  main.Notification:
    description: A notification generated by the system to alert users about important
      events
    properties:
      action_label:
        example: Gestisci Terapia
        type: string
      action_url:
        example: /hedgehogs/1
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      data:
        example: '{"days_overdue": 5}'
        type: string
      dismissed:
        example: false
        type: boolean
      expires_at:
        example: "2024-02-15T10:30:00Z"
        format: date-time
        type: string
      hedgehog:
        $ref: '#/definitions/main.Hedgehog'
      hedgehog_id:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      message:
        example: La terapia 'Antibiotico' per Spillo è scaduta il 15/01/2024
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/main.NotificationPriority'
        example: high
      read:
        example: false
        type: boolean
      therapy:
        $ref: '#/definitions/main.Therapy'
      therapy_id:
        example: 1
        type: integer
      title:
        example: Terapia Scaduta
        type: string
      type:
        allOf:
        - $ref: '#/definitions/main.NotificationType'
        example: therapy_expired
    type: object
  main.NotificationPriority:
    description: Priority level for notifications
    enum:
    - low
    - medium
    - high
    - critical
    type: string
    x-enum-comments:
      PriorityCritical: Urgent notifications requiring immediate attention
      PriorityHigh: Important notifications
      PriorityLow: Informational notifications
      PriorityMedium: Standard notifications
    x-enum-descriptions:
    - Informational notifications
    - Standard notifications
    - Important notifications
    - Urgent notifications requiring immediate attention
    x-enum-varnames:
    - PriorityLow
    - PriorityMedium
    - PriorityHigh
    - PriorityCritical
  main.NotificationSettings:
    description: Configuration settings for the notification system
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      email_address:
        example: admin@laninna.org
        type: string
      email_notifications_enabled:
        example: false
        type: boolean
      id:
        example: 1
        type: integer
      no_weighing_days:
        example: 7
        minimum: 1
        type: integer
      therapy_expired_enabled:
        example: true
        type: boolean
      therapy_expiring_days:
        example: 3
        minimum: 1
        type: integer
      updated_at:
        example: "2024-01-15T10:30:00Z"
        format: date-time
        type: string
      webhook_url:
        example: https://hooks.slack.com/services/xxx
        type: string
      weight_drop_days:
        example: 7
        minimum: 1
        type: integer
      weight_drop_threshold:
        example: 50
        minimum: 1
        type: number
      weight_stagnation_days:
        example: 14
        minimum: 1
        type: integer
    type: object
...
  main.TherapyAnalysis:
    properties:
      alert:
        type: boolean
      alert_reason:
        type: string
      days_active:
        type: integer
      days_until_end:
        type: integer
      end_date:
        type: string
      hedgehog_id:
        type: integer
      hedgehog_name:
        type: string
      start_date:
        type: string
      status:
        type: string
      therapy_id:
        type: integer
      therapy_name:
        type: string
    type: object
...
  main.WeightAnalysis:
    properties:
      alert:
        type: boolean
      alert_reason:
        type: string
      current_weight:
        type: number
      days_since_weigh:
        type: integer
      hedgehog_id:
        type: integer
      hedgehog_name:
        type: string
      last_weigh_date:
        type: string
      previous_weight:
        type: number
      trend:
        description: '"improving", "stable", "declining", "critical"'
        type: string
      weight_change:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for managing hedgehog rescue center operations
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: La Ninna - Hedgehog Management API
  version: "1.0"
paths:
  /analysis/therapy:
    get:
      consumes:
      - application/json
      description: Get analysis of active therapies with alerts for expiring or overdue
        therapies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.TherapyAnalysis'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get therapy analysis
      tags:
      - Analysis
  /analysis/weight:
    get:
      consumes:
      - application/json
      description: Get analysis of hedgehog weight trends and alerts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.WeightAnalysis'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get weight analysis
      tags:
      - Analysis
...
  /notification-settings:
    get:
      consumes:
      - application/json
      description: Get the current notification settings for the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.NotificationSettings'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notification settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Update the notification settings for the system
      parameters:
      - description: Notification settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/main.NotificationSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.NotificationSettings'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update notification settings
      tags:
      - Settings
  /notifications:
    get:
      consumes:
      - application/json
      description: Get list of notifications with optional filters
      parameters:
      - description: Filter by unread status
        in: query
        name: unread
        type: boolean
      - description: Filter by priority
        in: query
        name: priority
        type: string
      - description: Filter by notification type
        in: query
        name: type
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Notification'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notifications
      tags:
      - Notifications
  /notifications/{id}:
    delete:
      consumes:
      - application/json
      description: Dismiss a notification by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Dismiss notification
      tags:
      - Notifications
  /notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a notification as read by its ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Notification'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - Notifications
  /notifications/stats:
    get:
      consumes:
      - application/json
      description: Get statistics about notifications (counts by status, priority,
        and type)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get notification statistics
      tags:
      - Notifications
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"